extract some config from server.js
